+++ example/complete/controllers1/middleware_controllers/MiddlewareController.js
@@ -1,7 +1,12 @@
 module.exports = {
   get,
+  post
 };
 
+function post(req, res) {
+  res.send(`Request POST with middleware : ${req.middleware_name}`);
+}
+
 function get(req, res) {
-  res.send(`Request with middleware : ${req.middleware_name}`);
+  res.send(`Request GET with middleware : ${req.middleware_name}`);
 }