+++ index.js
@@ -1,1 +1,3 @@
 module.exports = require('./src/index');
+module.exports.MIDDLEWARE_LEVEL = require('./src/config/middleware');
+module.exports.METHOD = require('./src/config/methods');