+++ src/config/middleware.js
@@ -1,7 +1,6 @@
-module.exports.TYPE = {
-  APP: 'app-level',
-  ROUTER: 'router-level',
-  HANDLE: 'handling-level',
-  BUILTIN: 'built-in-level',
-  THIRD: 'third-party-level',
+module.exports = {
+  DEFAULT: 'app',
+  APP: 'app',
+  ROUTER: 'router',
+  ERROR: 'error',
 };