+++ src/lib/middlewares.js
@@ -1,5 +1,7 @@
 const middlewareExtractor = require('../extract/middlewares');
 const debug = require('debug')('ExpressImpRouter.routes.extract');
+const LEVEL = require('../config/middleware');
+const METHOD = require('../config/methods');
 
 let middlewares = [], isDebug = false;
 
@@ -9,13 +11,51 @@ module.exports = {
   get,
 };
 
-function get(route) {
-  return middlewares.filter((mid) => {
-    if(route) {
-      return mid.generated && route === mid.route;
+function get(level, route, method) {
+  let scopedMiddlewares = [];
+  for(let j in middlewares) {
+    let index = scopedMiddlewares.map(item => item.route).indexOf(middlewares[j].route);
+    if(index === -1) {
+      scopedMiddlewares.push({
+        ...middlewares[j],
+        middlewares: [middlewares[j].action],
+      });
+    } else {
+      scopedMiddlewares[index].middlewares.push(middlewares[j].action);
+    }
+    // if(middlewares[j].level === LEVEL.APP) {
+    //
+    // }
+  }
+
+  return scopedMiddlewares.filter((mid) => {
+    let filterMethod = true;
+    if(mid.method && method) {
+      if(Array.isArray(mid.method)) {
+        let isPresent = false;
+        for(let i in mid.method) {
+          if(mid.method[i].toUpperCase() === METHOD.ALL) {
+            isPresent = true;
+            break;
+          }
+
+          if(mid.method[i].toUpperCase() === method.toUpperCase()) {
+            isPresent = true;
+            break;
+          }
+        }
+        filterMethod = isPresent;
+      } else {
+        if(mid.method.toUpperCase() !== METHOD.ALL && mid.method.toUpperCase() !== method.toUpperCase()) {
+          filterMethod = false;
+        }
+      }
     }
 
-    return mid.generated;
+    return mid.generated &&
+      mid.level === level &&
+      (route ? mid.route === route : true) &&
+      filterMethod;
   });
 }